Removed files from `Medsmaker and merged with current files in `superbit-ngmix`. Should resolve #5.
