+++ Medsmaker/astro_config/psfex.config
@@ -26,10 +26,10 @@ PSFVAR_DEGREES  3               # Polynom degree for each group
 
 SAMPLE_AUTOSELECT  Y            # Automatically select the FWHM (Y/N) ?
 SAMPLEVAR_TYPE     SEEING       # File-to-file PSF variability: NONE or SEEING
-SAMPLE_FWHMRANGE   1.5,2.5     # Allowed FWHM range
-SAMPLE_VARIABILITY 0.5          # Allowed FWHM variability (1.0 = 100%)
-SAMPLE_MINSN       5            # Minimum S/N for a source to be used
-SAMPLE_MAXELLIP    0.9          # Maximum (A-B)/(A+B) for a source to be used
+SAMPLE_FWHMRANGE   1.6,2.5      # Allowed FWHM range
+SAMPLE_VARIABILITY 0.2          # Allowed FWHM variability (1.0 = 100%)
+SAMPLE_MINSN       8            # Minimum S/N for a source to be used
+SAMPLE_MAXELLIP    0.3          # Maximum (A-B)/(A+B) for a source to be used
 
 #----------------------------- Output catalogs -------------------------------
