+++ Medsmaker/astro_config/psfex.mock.config
@@ -6,7 +6,7 @@
 
 BASIS_TYPE      PIXEL_AUTO      # NONE, PIXEL, GAUSS-LAGUERRE or FILE
 BASIS_NUMBER    20              # Basis number or parameter
-PSF_SAMPLING    0.0               # Sampling step in pixel units (0.0 = auto)
+PSF_SAMPLING    0               # Sampling step in pixel units (0.0 = auto)
 PSF_ACCURACY    0.01            # Accuracy to expect from PSF "pixel" values
 PSF_SIZE        30,30           # Image size of the PSF model
 PSF_RECENTER    Y
@@ -20,16 +20,16 @@ PHOTFLUXERR_KEY FLUXERR_APER(1) # Catalogue parameter for photometric error
 
 PSFVAR_KEYS     X_IMAGE,Y_IMAGE # Catalogue or FITS (preceded by :) params
 PSFVAR_GROUPS   1,1             # Group tag for each context key
-PSFVAR_DEGREES  3               # Polynom degree for each group
+PSFVAR_DEGREES  2               # Polynom degree for each group
 
 #----------------------------- Sample selection ------------------------------
 
 SAMPLE_AUTOSELECT  Y            # Automatically select the FWHM (Y/N) ?
 SAMPLEVAR_TYPE     NONE         # File-to-file PSF variability: NONE or SEEING
-SAMPLE_FWHMRANGE   1.9,2.4      # Allowed FWHM range
-SAMPLE_VARIABILITY 0.08         # Allowed FWHM variability (1.0 = 100%)
-SAMPLE_MINSN       20          # Minimum S/N for a source to be used
-SAMPLE_MAXELLIP    0.12          # Maximum (A-B)/(A+B) for a source to be used
+SAMPLE_FWHMRANGE  1.7,2.05      # Allowed FWHM range (2.7,3.2)
+SAMPLE_VARIABILITY 0.6          # Allowed FWHM variability (1.0 = 100%)
+SAMPLE_MINSN       5            # Minimum S/N for a source to be used
+SAMPLE_MAXELLIP    0.7          # Maximum (A-B)/(A+B) for a source to be used
 
 #----------------------------- Output catalogs -------------------------------
