+++ Medsmaker/astro_config/sextractor.config
@@ -12,7 +12,7 @@ PARAMETERS_NAME  sextractor.param  # name of the file containing catalog content
 #------------------------------- Extraction ----------------------------------
 
 DETECT_TYPE      CCD            # CCD (linear) or PHOTO (with gamma correction)
-DETECT_MINAREA   3              # min. # of pixels above threshold
+DETECT_MINAREA   2              # min. # of pixels above threshold
 DETECT_MAXAREA   0              # max. # of pixels above threshold (0=unlimited)
 THRESH_TYPE      RELATIVE       # threshold type: RELATIVE (in sigmas)
                                 # or ABSOLUTE (in ADUs)
@@ -81,7 +81,7 @@ STARNNW_NAME     default.nnw    # Neural-Network_Weight table filename
 
 BACK_TYPE        AUTO           # AUTO or MANUAL
 BACK_VALUE       0.0            # Default background value in MANUAL mode
-BACK_SIZE        128             # Background mesh: <size> or <width>,<height>
+BACK_SIZE        256             # Background mesh: <size> or <width>,<height>
 BACK_FILTERSIZE  3              # Background filter: <size> or <width>,<height>
 
 BACKPHOTO_TYPE   LOCAL          # can be GLOBAL or LOCAL
