+++ Medsmaker/astro_config/swarp.config
@@ -10,7 +10,7 @@ HEADER_SUFFIX          .head           # Filename extension for additional heade
 
 #------------------------------- Input Weights --------------------------------
 
-WEIGHT_TYPE            MAP_WEIGHT      # BACKGROUND,MAP_RMS,MAP_VARIANCE
+WEIGHT_TYPE            NONE      # BACKGROUND,MAP_RMS,MAP_VARIANCE
                                        # or MAP_WEIGHT
 WEIGHT_SUFFIX          .weight.fits    # Suffix to use for weight-maps
 WEIGHT_IMAGE                           # Weightmap filename if suffix not used
