--- Medsmaker/astro_config_ccv/psfex.config
@@ -1,63 +1,0 @@
-# Default configuration file for PSFEx 3.17.1
-# EB 2016-06-28
-#
-
-#-------------------------------- PSF model ----------------------------------
-
-BASIS_TYPE      PIXEL_AUTO      # NONE, PIXEL, GAUSS-LAGUERRE or FILE
-BASIS_NUMBER    20              # Basis number or parameter
-PSF_SAMPLING    0.0             # Sampling step in pixel units (0.0 = auto)
-PSF_ACCURACY    0.01            # Accuracy to expect from PSF "pixel" values
-PSF_SIZE        30,30           # Image size of the PSF model
-
-#------------------------- Point source measurements -------------------------
-
-CENTER_KEYS     X_IMAGE,Y_IMAGE # Catalogue parameters for source pre-centering
-PHOTFLUX_KEY    FLUX_APER(1)    # Catalogue parameter for photometric norm.
-PHOTFLUXERR_KEY FLUXERR_APER(1) # Catalogue parameter for photometric error
-
-#----------------------------- PSF variability -------------------------------
-
-PSFVAR_KEYS     X_IMAGE,Y_IMAGE # Catalogue or FITS (preceded by :) params
-PSFVAR_GROUPS   1,1             # Group tag for each context key
-PSFVAR_DEGREES  3               # Polynom degree for each group
-
-#----------------------------- Sample selection ------------------------------
-
-SAMPLE_AUTOSELECT  Y            # Automatically select the FWHM (Y/N) ?
-SAMPLEVAR_TYPE     SEEING       # File-to-file PSF variability: NONE or SEEING
-SAMPLE_FWHMRANGE   1.6,2.5      # Allowed FWHM range
-SAMPLE_VARIABILITY 0.2          # Allowed FWHM variability (1.0 = 100%)
-SAMPLE_MINSN       8            # Minimum S/N for a source to be used
-SAMPLE_MAXELLIP    0.3          # Maximum (A-B)/(A+B) for a source to be used
-
-#----------------------------- Output catalogs -------------------------------
-
-OUTCAT_TYPE        FITS_LDAC      # NONE, ASCII_HEAD, ASCII, FITS_LDAC
-OUTCAT_NAME        psfex_out.cat  # Output catalog filename
-
-
-#------------------------------- Check-plots ----------------------------------
-
-CHECKPLOT_DEV       PNG         # NULL, XWIN, TK, PS, PSC, XFIG, PNG,
-                                # JPEG, AQT, PDF or SVG
-CHECKPLOT_TYPE      FWHM,ELLIPTICITY,COUNTS, COUNT_FRACTION, CHI2, RESIDUALS
-                                # or NONE
-CHECKPLOT_NAME      fwhm.png, ellipticity.png, counts.png, countfrac.png, chi2.png, resi.png
-
-#------------------------------ Check-Images ---------------------------------
-
-CHECKIMAGE_TYPE CHI,PROTOTYPES,SAMPLES,RESIDUALS,SNAPSHOTS
-                                # or MOFFAT,-MOFFAT,-SYMMETRICAL
-CHECKIMAGE_NAME chi.fits,proto.fits,samp.fits,resi.fits,snap.fits
-                                # Check-image filenames
-
-#----------------------------- Miscellaneous ---------------------------------
-
-PSF_DIR                         # Where to write PSFs (empty=same as input)
-VERBOSE_TYPE    FULL          # can be QUIET,NORMAL,LOG or FULL
-WRITE_XML       Y               # Write XML file (Y/N)?
-XML_NAME        psfex.xml       # Filename for XML output
-NTHREADS        0               # Number of simultaneous threads for
-                                # the SMP version of PSFEx
-                                # 0 = automatic
