--- Medsmaker/astro_config_ccv/sextractor.param
@@ -1,54 +1,0 @@
-ALPHAWIN_J2000
-DELTAWIN_J2000
-X_IMAGE
-Y_IMAGE
-NUMBER
-MAG_ISO
-MAGERR_ISO
-MAG_ISOCOR
-MAGERR_ISOCOR
-MAG_APER
-MAGERR_APER
-MAG_AUTO
-MAGERR_AUTO
-MAG_BEST
-MAGERR_BEST
-FLUX_AUTO
-FLUXERR_AUTO
-FLUX_BEST
-FLUXERR_BEST
-FLUX_ISO
-FLUXERR_ISO
-KRON_RADIUS
-BACKGROUND
-THRESHOLD
-FLUX_MAX
-ISOAREA_IMAGE
-X2_IMAGE
-Y2_IMAGE
-XY_IMAGE
-YMIN_IMAGE
-YMAX_IMAGE
-XMIN_IMAGE
-XMAX_IMAGE
-A_IMAGE
-B_IMAGE
-THETA_IMAGE
-MU_THRESHOLD
-MU_MAX
-ISOAREA_WORLD
-X_WORLD
-Y_WORLD
-A_WORLD
-B_WORLD
-THETA_WORLD
-FLAGS
-FWHM_IMAGE
-FWHM_WORLD
-ELONGATION
-CLASS_STAR
-FLUX_RADIUS(1)
-FLUX_APER(1)
-ELLIPTICITY
-VIGNET(32,32)
-SNR_WIN
