--- Medsmaker/meds-files/.gitattributes
@@ -1,3 +1,0 @@
-*/*.fits filter=lfs diff=lfs merge=lfs -text
-*.fits filter=lfs diff=lfs merge=lfs -text
-*.meds filter=lfs diff=lfs merge=lfs -text
