--- Medsmaker/meds-files/0.205PSFscale-oneIm.meds
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3f3f6f11ea558b98bb0207f2c7e6451d8dd40faeeb271f5fa839d91415a3307b
-size 14402880
