--- Medsmaker/meds-files/0.538PSFscale-oneIm.meds
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:24c536493501d270560bc3b33c908440c8ac2d5a1664062a00395b9afd6ec102
-size 16266240
