--- Medsmaker/meds-files/0.95FWHM-32x32PSF-0.206scale.meds
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9aa3000c55c4f16c35abd9179eda31b185b27785951d1ab23a24888f882c6d24
-size 64843200
