--- Medsmaker/meds-files/0.95FWHM-32x32PSF.meds
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a75487338aef932cb62912ff0831bf0ec9e30053fb19a7df672cfa80775a8d1b
-size 121104000
