--- Medsmaker/meds-files/quarantine/debug3.meds
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9cac44db88df7beb2c344b67f5df493bd3f9709063cde0a30c61278579d8102d
-size 70151040
