--- Medsmaker/medsconfig_mock.yaml
@@ -1,41 +1,0 @@
-# Input / Output
-# --------------
-
-# Path to the science input files (patterns accepted)
-science_files:  "../GalSim/output/mock_superbit_flight_jitter_only_oversampled_1x300.000?.fits"
-# The truth file to use for doing PSF star selection
-truth_file:     "../GalSim/output/truth_flight_jitter_only_oversampled_1x300.0001.dat"
-# Output name for the final MEDS file
-out_file:       "./output/mock.meds"
-
-# External download
-# -----------------
-
-# WGET location of the pre-generated mask file; supercedes master_dark and master_flat  
-mask_src:       "bee.astro.utoronto.ca/~bit/superbit-metcal-data/supermask.fits" 
-# Location to store / read the mask file
-mask_file:      "./data/supermask.fits"
-# WGET location for master dark and master flat; supercedes dark_files and flat_files
-mask_dark_src:  "bee.astro.utoronto.ca/~bit/superbit-metcal-data/master_dark_median.fits" 
-mask_flat_src:  "bee.astro.utoronto.ca/~bit/superbit-metcal-data/master_flat_median.fits" 
-# Location to store / read master dark / flat files
-mask_dark_file: "./data/master_dark_median.fits"
-mask_flat_file: "./data/master_flat_median.fits"
-
-# Working directories
-# -------------------
-working_dir:    "./output_debug"
-psfex_dir:      "./output_debug/psfex_output"
-
-# Tool configuration
-# -------------------
-swarp_config:   "./astro_config/swarp.config"
-psfex_config:   "./astro_config/psfex.config"
-sextractor_dir: "./astro_config/"
-
-# Raw data
-# --------
-flat_files:     null
-dark_files:     null
-bias_files:     null
-
