--- Medsmaker/superbit_meds.yaml
@@ -1,125 +1,0 @@
-# same config should work for cosmos and SN
-parspace: 'ngmix'
-
-hst_band: null
-
-# usually this is to skip the coadd
-skip_first_epoch: True
-
-weight_type: 'weight'
-
-trim_images:
-    min_size: 32
-    max_size: 64
-
-# gaussian aperture weighted fluxes
-gap:
-    # in arcsec
-    weight_fwhm: 2.5
-
-# ensure a box with specified radius is unmasked
-#cut_masked_center:
-#    radius_pixels: 3.0
-
-reject_outliers: True
-
-# for cosmos these are full meds.  Basically we expect the outlier
-# rejection to get everything except saturated stars and such,
-# which we will mask. So we don't do anything here.
-#
-# for the SN these are coadds which only BPM is set
-
-image_flagnames_to_mask: []
-
-# this is before adding additional masking, e.g. circular mask
-# does not apply for uberseg
-max_maskfrac: 0.45
-
-# minimum number of unmasked pixels to be used in an epoch.  this is only
-# checked if using uberseg, or in _set_weight, if the weight map is being
-# modified, e.g. circular mask
-min_npix: 8
-
-use_mask: True
-
-radius_column: "KRON_RADIUS"
-
-max_fof_size: 20
-
-fofs:
-    method: 'radius'
-    check_seg: false
-
-    # name in meds file, or catalog if method is catalog_radius
-    #radius_column: "iso_radius_arcsec"
-    radius_column: "KRON_RADIUS"
-
-    # factor to multiply radius this happens before clipping to [min_radius,
-    # max_radius] for a low threshold, the isoarea is basically covering all
-    # the observed flux, so mult of 1 makes sense
-    radius_mult: 1.0
-
-    # clip the radius in pixels=sqrt(isoarea_image/pi)
-    # 5 pixels is about 3 sigma for a 1'' FWHM gaussian
-    # this happens after the radius_mult is applied
-    min_radius_arcsec: 0.1
-
-    max_radius_arcsec: null
-    #max_radius_arcsec: 2.0
-
-    # This is added to the radius. This kind of padding makes sense for radii
-    # based on the iso area.  Padding happens after mult and clipping
-    padding_arcsec: 0.0
-
-    # arcsec
-    #extra_psf_fwhm_arcsec: 1.0
-    extra_psf_fwhm_arcsec: 0.0
-
-mof:
-
-    model: 'exp'
-    # use_input_guesses: True
-
-    # number of times to try the fit if it fails
-    ntry: 2
-
-    # for guesses
-    detband: 0
-
-    priors:
-        cen:
-            type: 'normal2d'
-            sigma: 0.001
-
-        g:
-            type: 'ba'
-            sigma: 0.2
-
-        T:
-            type: 'flat'
-            pars: [-1.0, 1.e+05]
-
-        flux:
-            type: 'flat'
-            pars: [-1000.0, 1.0e+09]
-
-        fracdev:
-            type: 'normal'
-            mean: 0.5
-            sigma: 0.1
-            bounds: [0.0, 1.0]
-
-    psf:
-        ntry: 4
-
-        model: 'em3' # Expectations Maximization; em3 means 3 Gaussian models
-
-        em_pars:
-            maxiter: 2000
-            tol: 1.0e-4
-
-    # higher tolerance for groups and SOF failures
-    lm_pars:
-        maxfev: 4000
-        ftol: 5.0e-5
-        xtol: 5.0e-5
